Move getUnusedSlug/etc out of shared 'lib', into server-only directory
