+++ packages/lesswrong/server/authentication/auth0Accounts.ts
@@ -1,5 +1,6 @@
 import { Profile } from "passport-auth0";
-import { slugify, Utils } from "../../lib/vulcan-lib/utils";
+import { slugify } from "../../lib/vulcan-lib/utils";
+import { getUnusedSlugByCollectionName } from '../utils/slugUtils';
 
 export async function userFromAuth0Profile(profile: Profile): Promise<Partial<DbUser>> {
   const email = profile.emails?.[0].value
@@ -18,7 +19,7 @@ export async function userFromAuth0Profile(profile: Profile): Promise<Partial<Db
     services: {
       auth0: profile
     },
-    username: await Utils.getUnusedSlugByCollectionName("Users", slugify(displayName)),
+    username: await getUnusedSlugByCollectionName("Users", slugify(displayName)),
     displayName: displayName,
     usernameUnset: true
   }
