+++ packages/lesswrong/server/resolvers/postResolvers.ts
@@ -11,10 +11,23 @@ import { matchSideComments } from '../sideComments';
 import { captureException } from '@sentry/core';
 import { getToCforPost } from '../tableOfContents';
 import { getDefaultViewSelector } from '../../lib/utils/viewUtils';
+import { getUnusedSlugByCollectionName } from '../utils/slugUtils';
+import { slugify } from '../../lib/vulcan-lib/utils';
 import keyBy from 'lodash/keyBy';
 import GraphQLJSON from 'graphql-type-json';
 
 augmentFieldsDict(Posts, {
+  slug: {
+    onInsert: async (post) => {
+      return await getUnusedSlugByCollectionName("Posts", slugify(post.title))
+    },
+    onEdit: async (modifier, post) => {
+      if (modifier.$set.title) {
+        return await getUnusedSlugByCollectionName("Posts", slugify(modifier.$set.title), false, post._id)
+      }
+    }
+  },
+
   // Compute a denormalized start/end time for events, accounting for the
   // timezone the event's location is in. This is subtly wrong: it computes a
   // correct timestamp, but then the timezone part of that timezone gets lost
